Edit: 파일 변경
JoinMemberShip 컴포넌트 내용 -> SetMemberShip 컴포넌트로 이동
