+++ src/pages/auth/JoinMemberShip/index.style.js
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-export const Main = styled.div`
+export const MainLayOut = styled.div`
   display: flex;
   flex-direction: column;
   align-items: center;
